fix types of price, ebook_price and that date_update is nullable

+++ Services/Courses/CoursesService.API/DTOs/CourseDetailDto.cs
@@ -9,8 +9,8 @@ public class CourseDetailDto
     public string Name { get; set; } = string.Empty;
     public string Slug { get; set; } = string.Empty;
     public string? Description { get; set; }
-    public decimal Price { get; set; }
-    public decimal? EbookPrice { get; set; }
+    public float Price { get; set; }
+    public float? EbookPrice { get; set; }
     public Guid? Thumbnail { get; set; }
     public Guid? EbookPdf { get; set; }
     public string? Level { get; set; }
